docs/requirements.txt
python-docx 
python-pptx 
pillow 
pandas 
PyPDF2 
PyMuPDF
openpyxl